Move tests and fixtures into separate files
