+++ tests/unittests/portadmin/napalm/juniper_test.py
@@ -19,16 +19,11 @@ import pytest
 from unittest.mock import Mock, patch
 
 from jnpr.junos.exception import RpcError
-from lxml import etree
 
-from nav.enterprise.ids import VENDOR_ID_RESERVED, VENDOR_ID_JUNIPER_NETWORKS_INC
-from nav.models import manage
+from nav.enterprise.ids import VENDOR_ID_RESERVED
 from nav.portadmin.handlers import (
     DeviceNotConfigurableError,
     ProtocolError,
-    POEStateNotSupportedError,
-    POENotSupportedError,
-    XMLParseError,
 )
 from nav.portadmin.napalm.juniper import wrap_unhandled_rpc_errors, Juniper
 
@@ -92,195 +87,3 @@ class TestJuniper:
 
         m = MockedJuniperHandler(Mock())
         assert len(m.get_netbox_vlans()) == 1
-
-
-class TestJuniperPoe:
-    def test_returns_correct_state_options(self, handler_mock):
-        state_options = handler_mock.get_poe_state_options()
-        assert Juniper.POE_ENABLED in state_options
-        assert Juniper.POE_DISABLED in state_options
-
-    def test_state_converter_returns_correct_states(self, handler_mock):
-        assert handler_mock._poe_string_to_state("enabled") == Juniper.POE_ENABLED
-        assert handler_mock._poe_string_to_state("disabled") == Juniper.POE_DISABLED
-
-    def test_state_converter_raises_error_for_invalid_states(self, handler_mock):
-        with pytest.raises(POEStateNotSupportedError):
-            handler_mock._poe_string_to_state("invalid_state")
-
-    def test_get_poe_states_uses_interfaces_from_db_if_input_is_none(
-        self, handler_mock, xml_bulk
-    ):
-        expected_interfaces = handler_mock.netbox.interfaces
-        handler_mock._get_all_poe_interface_information = Mock(return_value=xml_bulk)
-        return_dict = handler_mock.get_poe_states()
-        for interface in expected_interfaces:
-            assert interface.ifindex in return_dict
-
-    def test_get_poe_states_uses_interfaces_from_db_if_input_is_empty(
-        self, handler_mock, xml_bulk
-    ):
-        expected_interfaces = handler_mock.netbox.interfaces
-        handler_mock._get_all_poe_interface_information = Mock(return_value=xml_bulk)
-        return_dict = handler_mock.get_poe_states([])
-        for interface in expected_interfaces:
-            assert interface.ifindex in return_dict
-
-
-class TestGetSinglePoeState:
-    def test_returns_correct_state_for_interface_that_exists_in_xml_response(
-        self, handler_mock, xml, interface1_mock
-    ):
-        handler_mock._get_poe_interface_information = Mock(return_value=xml)
-        state = handler_mock._get_single_poe_state(interface1_mock)
-        assert state == Juniper.POE_ENABLED
-
-    def test_raises_exception_if_no_interfaces_in_xml(
-        self, handler_mock, interface1_mock, xml_empty
-    ):
-        handler_mock._get_poe_interface_information = Mock(return_value=xml_empty)
-        with pytest.raises(POENotSupportedError):
-            handler_mock._get_single_poe_state(interface1_mock)
-
-    def test_raises_exception_if_multiple_interfaces_in_xml(
-        self, handler_mock, interface1_mock, xml_bulk_wrong_format
-    ):
-        handler_mock._get_poe_interface_information = Mock(
-            return_value=xml_bulk_wrong_format
-        )
-        with pytest.raises(XMLParseError):
-            handler_mock._get_single_poe_state(interface1_mock)
-
-
-class TestGetPoeStatesBulk:
-    def test_returns_correct_states(
-        self, handler_mock, xml_bulk, interface1_mock, interface2_mock
-    ):
-        handler_mock._get_all_poe_interface_information = Mock(return_value=xml_bulk)
-        states = handler_mock._get_poe_states_bulk([interface1_mock, interface2_mock])
-        assert states[interface1_mock.ifindex] == Juniper.POE_ENABLED
-        assert states[interface2_mock.ifindex] == Juniper.POE_DISABLED
-
-    def test_maps_interface_to_none_if_poe_not_supported(self, handler_mock, xml_bulk):
-        handler_mock._get_all_poe_interface_information = Mock(return_value=xml_bulk)
-        if_mock = Mock()
-        if_mock.ifname == "random_if"
-        if_mock.ifindex = 0
-        states = handler_mock._get_poe_states_bulk([if_mock])
-        assert states[if_mock.ifindex] is None
-
-    def test_returns_none_values_if_no_interfaces_in_xml(
-        self, handler_mock, interface1_mock, interface2_mock, xml_empty
-    ):
-        handler_mock._get_all_poe_interface_information = Mock(return_value=xml_empty)
-        return_dict = handler_mock._get_poe_states_bulk(
-            [interface1_mock, interface2_mock]
-        )
-        assert return_dict[interface1_mock.ifindex] is None
-        assert return_dict[interface2_mock.ifindex] is None
-
-
-@pytest.fixture()
-def netbox_mock(interface1_mock, interface2_mock):
-    """Create netbox model mock object"""
-    netbox = Mock()
-    netbox.ip = '10.0.0.1'
-    netbox.type.get_enterprise_id.return_value = VENDOR_ID_JUNIPER_NETWORKS_INC
-    netbox.interfaces = [interface1_mock, interface2_mock]
-    yield netbox
-
-
-@pytest.fixture()
-def profile_mock():
-    """Create management profile model mock object"""
-    profile = Mock()
-    profile.protocol = manage.ManagementProfile.PROTOCOL_NAPALM
-    profile.PROTOCOL_NAPALM = manage.ManagementProfile.PROTOCOL_NAPALM
-    profile.configuration = {"driver": "mock"}
-    yield profile
-
-
-@pytest.fixture()
-def handler_mock(netbox_mock, profile_mock):
-    """Create management handler mock object"""
-    juniper = Juniper(netbox=netbox_mock)
-    juniper._profile = profile_mock
-    yield juniper
-
-
-@pytest.fixture()
-def xml(interface1_mock):
-    """Creates a ElementTree containing poe information for one interface"""
-    tree_string = f"""
-        <poe>
-            <interface-information-detail>
-                <interface-name-detail>{interface1_mock.ifname}</interface-name-detail>
-                <interface-enabled-detail>Enabled</interface-enabled-detail>
-            </interface-information-detail>
-        </poe>"""
-    tree = etree.fromstring(tree_string)
-    yield tree
-
-
-@pytest.fixture()
-def xml_bulk_wrong_format(interface1_mock, interface2_mock):
-    """Creates a ElementTree with the format meant for a single interface in the response,
-    but it contains poe information for two interfaces
-    """
-    tree_string = f"""
-        <poe>
-            <interface-information-detail>
-                <interface-name-detail>{interface1_mock.ifname}</interface-name-detail>
-                <interface-enabled-detail>Enabled</interface-enabled-detail>
-            </interface-information-detail>
-            <interface-information-detail>
-                <interface-name-detail>{interface2_mock.ifname}</interface-name-detail>
-                <interface-enabled-detail>Enabled</interface-enabled-detail>
-            </interface-information-detail>
-        </poe>"""
-    tree = etree.fromstring(tree_string)
-    yield tree
-
-
-@pytest.fixture()
-def xml_bulk(interface1_mock, interface2_mock):
-    """Creates a ElementTree containing poe information for two interfaces"""
-    tree_string = f"""
-        <poe>
-            <interface-information>
-                <interface-name>{interface1_mock.ifname}</interface-name>
-                <interface-enabled>Enabled</interface-enabled>
-            </interface-information>
-            <interface-information>
-                <interface-name>{interface2_mock.ifname}</interface-name>
-                <interface-enabled>Disabled</interface-enabled>
-            </interface-information>
-        </poe>"""
-    tree = etree.fromstring(tree_string)
-    yield tree
-
-
-@pytest.fixture()
-def xml_empty():
-    """Creates a ElementTree containing no poe state for any interface"""
-    tree_string = f"""
-        <poe>
-        </poe>"""
-    tree = etree.fromstring(tree_string)
-    yield tree
-
-
-@pytest.fixture()
-def interface1_mock():
-    interface = Mock()
-    interface.ifname = "ge-0/0/1"
-    interface.ifindex = 1
-    yield interface
-
-
-@pytest.fixture()
-def interface2_mock():
-    interface = Mock()
-    interface.ifname = "ge-0/0/2"
-    interface.ifindex = 2
-    yield interface
